+++ src/lexer/test.rs
@@ -58,3 +58,21 @@ fn test_numerical_tokens() {
         assert_token(t, Literal(Imaginary), Some(t));
     }
 }
+
+#[test]
+fn test_text_literals() {
+    use super::TokenKind::Literal;
+    use super::Literal::*;
+
+    assert_token("'a'", Literal(Rune), Some("a"));
+    assert_token("'\\n'", Literal(Rune), Some("\\n"));
+    assert_token("'\\''", Literal(Rune), Some("\\'"));
+
+    assert_token("\"Hello!\"", Literal(Str), Some("Hello!"));
+    assert_token("\"\\n\\n\"", Literal(Str), Some("\\n\\n"));
+    assert_token("\"\\\"\"", Literal(Str), Some("\\\""));
+
+    assert_token("`Hello!`", Literal(StrRaw), Some("Hello!"));
+    assert_token("`\\n\\n`", Literal(StrRaw), Some("\\n\\n"));
+    assert_token("`\\\"`", Literal(StrRaw), Some("\\\""));
+}
